object_lambda_benchmark/microbenchmarks/coldstart/python/requirements.txt
boto3
Cython
grpcio==1.44.0
grpcio-tools==1.44.0
google
protobuf
